manual fixes from static analysis findings

+++ AprgAlgorithm/AprgAlgorithm/tst/AlgorithmTests/Sort/RadixSorter/LeastSignificantDigitSorter_unit.cpp
@@ -44,7 +44,7 @@ StringsSorter::GetDigitAtFunction getCharacterAtForString = [](string const& val
                                                                int const mostSignificantDigitIndex) -> int {
     int digitValue{};
     if (mostSignificantDigitIndex < static_cast<int>(value.length())) {
-        digitValue = value[mostSignificantDigitIndex];
+        digitValue = static_cast<int>(value[mostSignificantDigitIndex]);
     }
     return digitValue;
 };
